Move two functions to a new module: polygon_utils
(index, lib/polygon_utils).js
	-Move two polygon computational functions, `getBoundingBox()`
	and `simplifyCoords()`, to a new module called `polygon_utils`.
