+++ index.js
@@ -3,8 +3,9 @@
 var shapefileStream = require( 'shapefile-stream' );
 var rbush = require( 'rbush' );
 var pointInPolygon = require( 'point-in-polygon' );
-var simplify = require( 'simplify-js' );
 var through = require( 'through2' );
+
+var polygonUtils = require( './lib/polygon_utils' );
 
 function HierarchyLookup( rtree, polygons ){
   this.rtree = rtree;
@@ -27,54 +28,13 @@ HierarchyLookup.prototype.search = function search( lat, lon ){
   }
 };
 
-function getBoundingBox( poly ){
-  var firstPt = poly.shift();
-  var bbox = [
-    firstPt[ 0 ], firstPt[ 1 ],
-    firstPt[ 0 ], firstPt[ 1 ]
-  ];
-
-  poly.forEach( function identifyBounds( pt ){
-    var x = pt[ 0 ];
-    if( x < bbox[ 0 ] ){
-      bbox[ 0 ] = x;
-    }
-    else if( x > bbox[ 2 ] ){
-      bbox[ 2 ] = x;
-    }
-
-    var y = pt[ 1 ];
-    if( y < bbox[ 1 ] ){
-      bbox[ 1 ] = y;
-    }
-    else if( y > bbox[ 3 ] ){
-      bbox[ 3 ] = y;
-    }
-  });
-
-  return bbox;
-}
-
-function simplifyCoords( coords ){
-  var pts = coords.map( function mapToSimplifyFmt( pt ){
-    return { x: pt[ 0 ], y: pt[ 1 ] };
-  });
-
-  var simplificationRate = 0.0026;
-  var simplified = simplify( pts, simplificationRate, true );
-
-  return simplified.map( function mapToGeoJsonFmt( pt ){
-    return [ pt.x, pt.y ];
-  });
-}
-
 function load(path, desiredProps, cb){
   var bboxes = [];
   var polygons = [];
   var id = 0;
 
   function indexPolygon( props, coords ){
-    var simplified = simplifyCoords( coords );
+    var simplified = polygonUtils.simplifyCoords( coords );
     var propSubset = {};
     desiredProps.forEach( function ( propName ){
       propSubset[ propName ] = props[ propName ];
@@ -85,7 +45,7 @@ function load(path, desiredProps, cb){
       properties: propSubset
     });
 
-    var bbox = getBoundingBox( simplified );
+    var bbox = polygonUtils.getBoundingBox( simplified );
     bbox.id = id++;
     bboxes.push(bbox);
   }
